backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
aiosqlite==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.3.1
pytest-asyncio==0.23.7
anyio==4.4.0

